Fix int v long format specifier conflicts Original rev  : rc@rc0.org.uk--2004-mairix/mairix--dev--0.15--patch-22

+++ rfc822.c
@@ -711,7 +711,7 @@ static void do_attachment(char *start, char *after_end, struct attachment *atts)
   if (body_start > after_end) {
     /* This is a (maliciously?) b0rken attachment, e.g. maybe empty */
     if (verbose) {
-      fprintf(stderr, "This message contains an invalid attachment, length=%d bytes\n", after_end - start);
+      fprintf(stderr, "This message contains an invalid attachment, length=%d bytes\n", (int)(after_end - start));
     }
   } else {
     body_len = after_end - body_start;
